Rename onboarding.ts to admin.ts
