+++ src/vues/components/onboarding/FullFolderStatus.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import type { PathInfoResponse } from '@/scripts/types/api/onboarding.ts'
+import type { PathInfoResponse } from '@/scripts/types/api/admin.js'
 import { prettyBytes } from '@/scripts/utils.ts'
 
 defineProps<{
